requirements.txt
pandas
numpy
matplotlib
seaborn
google-genai
streamlit
python-dotenv